Remove other M2 related code and update unit tests.
