+++ WooCommerce/WooCommerceTests/ViewRelated/Products/Edit Product/ProductFormActionsFactory+EditProductsM3Tests.swift
@@ -14,7 +14,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -42,7 +41,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -70,7 +68,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -97,7 +94,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -124,7 +120,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -148,7 +143,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -169,7 +163,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
@@ -190,7 +183,6 @@ final class ProductFormActionsFactory_EditProductsM3Tests: XCTestCase {
 
         // Action
         let factory = ProductFormActionsFactory(product: model,
-                                                isEditProductsRelease2Enabled: true,
                                                 isEditProductsRelease3Enabled: true)
 
         // Assert
