+++ WooCommerce/WooCommerceTests/ViewRelated/Products/Edit Product/ProductFormViewModel+UpdatesTests.swift
@@ -13,7 +13,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -31,7 +30,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -49,7 +47,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -80,7 +77,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -113,7 +109,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -146,7 +141,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -170,7 +164,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: true)
 
         // Action
@@ -192,7 +185,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: true)
 
         // Action
@@ -213,7 +205,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -231,7 +222,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -272,7 +262,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -290,7 +279,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
@@ -310,7 +298,6 @@ final class ProductFormViewModel_UpdatesTests: XCTestCase {
         let productImageActionHandler = ProductImageActionHandler(siteID: 0, product: model)
         let viewModel = ProductFormViewModel(product: model,
                                              productImageActionHandler: productImageActionHandler,
-                                             isEditProductsRelease2Enabled: true,
                                              isEditProductsRelease3Enabled: false)
 
         // Action
